+++ src/main/java/com/dnd/snappy/controller/v1/auth/interceptor/MeetingParticipationInterceptor.java
@@ -1,8 +1,8 @@
 package com.dnd.snappy.controller.v1.auth.interceptor;
 
 import com.dnd.snappy.common.error.exception.BusinessException;
-import com.dnd.snappy.controller.v1.auth.JwtTokenExtractor;
-import com.dnd.snappy.controller.v1.auth.PathVariableExtractor;
+import com.dnd.snappy.domain.auth.service.JwtTokenExtractor;
+import com.dnd.snappy.domain.auth.service.PathVariableExtractor;
 import com.dnd.snappy.domain.participant.exception.ParticipantErrorCode;
 import com.dnd.snappy.domain.participant.repository.ParticipantRepository;
 import com.dnd.snappy.domain.token.service.TokenProvider;
@@ -26,26 +26,26 @@ public class MeetingParticipationInterceptor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        if (isMeetingParticipationUrl(request)) {
+        if (isMeetingParticipationRequest(request)) {
             return true;
         }
 
-        Long meetingId = pathVariableExtractor.extractMeetingId(request);
+        final Long meetingId = pathVariableExtractor.extractMeetingId(request);
         final String token = jwtTokenExtractor.extractToken(request, meetingId, TokenType.ACCESS_TOKEN);
-        Long participantId = tokenProvider.extractPayload(token);
+        final Long participantId = tokenProvider.extractPayload(token);
         validationParticipantInMeeting(participantId, meetingId);
 
         return true;
     }
 
-    private boolean isMeetingParticipationUrl(HttpServletRequest request) {
-        String requestUri = request.getRequestURI();
-        return requestUri.matches(PARTICIPATION_URL_PATTERN) && request.getMethod().equalsIgnoreCase("POST");
-    }
-
-    public void validationParticipantInMeeting(Long participantId, Long meetingId) {
+    private void validationParticipantInMeeting(Long participantId, Long meetingId) {
         if(!participantRepository.existsByIdAndMeetingId(participantId, meetingId)) {
             throw new BusinessException(ParticipantErrorCode.NOT_PARTICIPATING_MEETING);
         }
+    }
+
+    private boolean isMeetingParticipationRequest(HttpServletRequest request) {
+        String requestUri = request.getRequestURI();
+        return requestUri.matches(PARTICIPATION_URL_PATTERN) && request.getMethod().equalsIgnoreCase("POST");
     }
 }
