+++ src/main/java/com/dnd/snappy/controller/v1/auth/resolver/AuthInfo.java
@@ -1,0 +1,6 @@
+package com.dnd.snappy.controller.v1.auth.resolver;
+
+public record AuthInfo(
+        Long participantId
+) {
+}
