+++ src/main/java/com/dnd/snappy/controller/v1/participant/ParticipantController.java
@@ -1,7 +1,7 @@
 package com.dnd.snappy.controller.v1.participant;
 
 import com.dnd.snappy.common.dto.ResponseDto;
-import com.dnd.snappy.controller.v1.auth.AuthCookieManager;
+import com.dnd.snappy.domain.auth.service.AuthCookieManager;
 import com.dnd.snappy.controller.v1.participant.request.ParticipationRequest;
 import com.dnd.snappy.controller.v1.participant.response.ParticipationResponse;
 import com.dnd.snappy.domain.participant.service.ParticipationService;
