--- src/main/java/com/dnd/snappy/controller/v1/participant/resolver/AuthInfo.java
@@ -1,6 +1,0 @@
-package com.dnd.snappy.controller.v1.participant.resolver;
-
-public record AuthInfo(
-        Long participantId
-) {
-}
