+++ src/main/java/com/dnd/snappy/domain/auth/service/AuthCookieManager.java
@@ -1,7 +1,5 @@
-package com.dnd.snappy.controller.v1.auth;
+package com.dnd.snappy.domain.auth.service;
 
-import com.dnd.snappy.common.error.CommonErrorCode;
-import com.dnd.snappy.common.error.exception.BusinessException;
 import com.dnd.snappy.domain.token.service.TokenType;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
